add get_path_to_leaf
